refactor: Input 타입 분리
